Move celery task under django module
